feat: add contact form functionality with rate limiting and email notifications
added: sections folder
- every section in there and with import in the home = easy editing
