+++ app/Mail/ContactMessage.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Mail;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+
+class ContactMessage extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    public array $data;
+
+    public function __construct(array $data) { $this->data = $data; }
+
+    public function build()
+    {
+        return $this->subject($this->data['subject'] ?: 'New contact message')
+            ->markdown('emails.contact', ['data' => $this->data]);
+    }
+}
