+++ resources/js/pages/Sections/About.vue
@@ -1,0 +1,94 @@
+<script setup lang="ts">
+
+</script>
+<template>
+    <section id="about" class="px-8 py-24">
+        <div class="mx-auto max-w-6xl">
+            <div class="flex items-center pb-8 lg:justify-center">
+                <h1 class="text-4xl font-bold text-black dark:text-white">About me<span class="text-primary">.</span></h1>
+            </div>
+            <div class="grid grid-cols-1 gap-32 md:grid-cols-2">
+                <div class="flex flex-col gap-3 text-black dark:text-white">
+                    <h2 class="text-2xl font-bold">Passionate Software Developer Student</h2>
+                    <p>
+                        Hey There! I'm Daniel Terpstra, 19 years old and a passionate student software developer based in the Netherlands. I
+                        specialize in crafting beautiful and functional web applications that provide seamless user experiences, with a strong
+                        foundation in both frontend and backend. Right now I'm focused on learning
+                        <a target="_blank" href="https://laravel.com/" class="cursor-pointer font-bold transition hover:text-[#FF2D20]"
+                        >Laravel</a
+                        >
+                        and
+                        <a target="_blank" href="https://filamentphp.com/" class="cursor-pointer font-bold transition hover:text-[#FDAE4B]"
+                        >Filament</a
+                        >
+                        for backend development. On the frontend I’m focused on the
+                        <a target="_blank" href="https://tallstack.dev/" class="cursor-pointer font-bold transition hover:text-[#38bdf8]"
+                        >TALL stack</a
+                        >, which combines
+                        <a target="_blank" href="https://tailwindcss.com/" class="cursor-pointer font-bold transition hover:text-[#06B6D4]"
+                        >TailwindCSS</a
+                        >,
+                        <a target="_blank" href="https://alpinejs.dev/" class="cursor-pointer font-bold transition hover:text-[#8BC0D0]"
+                        >AlpineJS</a
+                        >,
+                        <a target="_blank" href="https://laravel.com/" class="cursor-pointer font-bold transition hover:text-[#FF2D20]"
+                        >Laravel</a
+                        >
+                        and
+                        <a target="_blank" href="https://laravel-livewire.com/" class="cursor-pointer font-bold transition hover:text-[#4E56A6]"
+                        >Livewire</a
+                        >. I’m also experienced with
+                        <a target="_blank" href="https://vuejs.org/" class="cursor-pointer font-bold transition hover:text-[#4FC08D]">Vue.js</a>,
+                        using
+                        <a target="_blank" href="https://primevue.org/" class="cursor-pointer font-bold transition hover:text-[#41B883]"
+                        >PrimeVue</a
+                        >
+                        and
+                        <a target="_blank" href="https://headlessui.dev/" class="cursor-pointer font-bold transition hover:text-[#66E3FF]"
+                        >HeadlessUI</a
+                        >
+                        for building dynamic user interfaces.
+                    </p>
+                    <div class="flex flex-row gap-3">
+                        <a
+                            class="rounded-xl bg-primary px-5 py-3 text-sm font-semibold text-white shadow-lg shadow-violet-600/30 transition duration-300 hover:bg-primary/80"
+                            href="#contact"
+                        >Lets Connect! 🚀</a
+                        >
+                    </div>
+                </div>
+                <div>
+                    <img class="w-full rounded-lg shadow-lg" src="images/placeholder.svg" alt="" />
+                </div>
+            </div>
+            <div class="relative pt-18">
+                <!-- background glow -->
+                <div class="absolute inset-0 -z-10 mt-16 rounded-2xl bg-primary/30 blur-3xl lg:blur-2xl"></div>
+
+                <!-- your stats grid -->
+                <div class="grid grid-cols-1 items-center gap-3 text-black md:grid-cols-2 lg:grid-cols-4 dark:text-zinc-200">
+                    <div class="flex flex-col items-center gap-2">
+                        <h1 class="text-3xl font-bold text-primary">3+ Years</h1>
+                        <p class="text-md">Experience in Software Development</p>
+                    </div>
+
+                    <div class="flex flex-col items-center gap-2">
+                        <h1 class="text-3xl font-bold text-primary">8+ Skills</h1>
+                        <p class="text-md">Frontend and Backend Technologies</p>
+                    </div>
+
+                    <div class="flex flex-col items-center gap-2">
+                        <h1 class="text-3xl font-bold text-primary">Focused On</h1>
+                        <p class="text-md">Frontend, backend & Fullstack skills</p>
+                    </div>
+
+                    <div class="flex flex-col items-center gap-2">
+                        <h1 class="text-3xl font-bold text-primary">Passionate About</h1>
+                        <p class="text-md">Crafting Clean & Modern Applications</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+
+</template>
