+++ resources/js/pages/Sections/Timeline.vue
@@ -1,0 +1,188 @@
+<script setup lang="ts">
+
+</script>
+<template>
+    <!-- DESKTOP timeiline-->
+    <section id="timeline" class="hidden px-8 py-24 md:block">
+        <div class="mx-auto max-w-6xl">
+            <div class="flex justify-center pb-12">
+                <h1 class="text-4xl font-bold text-black dark:text-white">Education<span class="text-purple-500">.</span></h1>
+            </div>
+
+            <div class="relative">
+                <!-- Vertical line -->
+                <div class="absolute top-0 left-1/2 h-full w-px -translate-x-1/2 bg-purple-700/40"></div>
+
+                <!-- ITEM 1 -->
+                <div class="relative mb-12 flex items-center justify-between">
+                    <div class="group w-5/12 text-right">
+                        <div
+                            class="cursor-default rounded-lg border border-primary/10 p-6 shadow-lg backdrop-blur-sm transition duration-300 hover:scale-105 hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-2 font-bold text-black dark:text-white">
+                                Bonhoeffer College locatie Van der Waalslaan -
+                                <span class="text-sm font-medium text-zinc-400">4years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                I completed my secondary education at Bonhoeffer College, where I got my MAVO diploma in 4 years. During the COVID
+                                period I spent a lot of time behind my computer doing online classes. That’s also when I found out I really
+                                enjoyed working with computers — I started taking them apart and putting them back together. That curiosity was
+                                the start of my journey into software development.
+                            </p>
+                        </div>
+                    </div>
+
+                    <!-- Dot -->
+                    <div class="relative flex items-center justify-center">
+                        <div class="absolute -z-10 h-16 w-16 rounded-full bg-purple-600/40 blur-xl"></div>
+                        <div class="h-6 w-6 rounded-full bg-purple-600 shadow-md"></div>
+                    </div>
+                    <div class="w-5/12"></div>
+                </div>
+
+                <!-- ITEM 2 -->
+                <div class="relative mb-12 flex items-center justify-between">
+                    <div class="w-5/12"></div>
+
+                    <!-- Dot -->
+                    <div class="relative flex items-center justify-center">
+                        <div class="absolute -z-10 h-16 w-16 rounded-full bg-purple-600/40 blur-xl"></div>
+                        <div class="h-6 w-6 rounded-full bg-purple-600 shadow-md"></div>
+                    </div>
+
+                    <div class="group w-5/12">
+                        <div
+                            class="cursor-default rounded-lg border border-primary/10 p-6 shadow-lg backdrop-blur-sm transition duration-300 hover:scale-105 hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-2 font-bold text-black dark:text-white">
+                                MBO Software Developer Study Enschede -
+                                <span class="text-sm font-medium text-zinc-400">2years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                This is really where it all began for me. For the intake I had to make a small console app in C#. Back then I had
+                                no clue what I was doing — I had to use things I’d never even heard of before. Still, I managed to build it and I
+                                was proud of myself. The study was more focused on game development, but after my first internship at C-o-d-e-s in
+                                Almelo, where I learned Laravel and Vue.js, I knew web development was the path I wanted to take.
+                            </p>
+                        </div>
+                    </div>
+                </div>
+
+                <!-- ITEM 3 -->
+                <div class="relative mb-12 flex items-center justify-between">
+                    <div class="group w-5/12 text-right">
+                        <div
+                            class="cursor-default rounded-lg border border-primary/10 p-6 shadow-lg backdrop-blur-sm transition duration-300 hover:scale-105 hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-2 font-bold text-black dark:text-white">
+                                MBO Software Developer Study Hengelo -
+                                <span class="text-sm font-medium text-zinc-400">2years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                After 2 years in Enschede I decided to switch from game development to web development. Together with 4 friends I
+                                moved to the Hengelo location. At first it wasn’t that easy, since we jumped straight into the third year and
+                                missed a lot of things the others already learned. But with some hard work we caught up quickly. Right now I’m in
+                                my fourth year doing my internship at Cube in Oldenzaal, where I’m learning a lot every day. After this, it’s time
+                                for my final exams.
+                            </p>
+                        </div>
+                    </div>
+
+                    <!-- Dot -->
+                    <div class="relative flex items-center justify-center">
+                        <div class="absolute -z-10 h-16 w-16 rounded-full bg-purple-600/40 blur-xl"></div>
+                        <div class="h-6 w-6 rounded-full bg-purple-600 shadow-md"></div>
+                    </div>
+
+                    <div class="w-5/12"></div>
+                </div>
+            </div>
+        </div>
+    </section>
+    <!-- MOBILE timeline-->
+    <section id="timeline" class="px-8 py-16 md:hidden">
+        <div class="mx-auto max-w-xl">
+            <div class="flex pb-8">
+                <h1 class="text-4xl font-bold text-black dark:text-white">Education<span class="text-purple-500">.</span></h1>
+            </div>
+
+            <div class="relative">
+                <!-- Linker lijn -->
+                <div class="absolute top-0 bottom-0 left-2 w-px bg-purple-700/40"></div>
+                <div class="space-y-10">
+                    <!-- ITEM 1 -->
+                    <div class="relative pl-12">
+                        <!-- Dot + glow -->
+                        <div class="absolute top-2 left-4 -translate-x-1/2">
+                            <div class="absolute -z-10 h-10 w-10 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600/35 blur-lg"></div>
+                            <div class="h-4 w-4 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600 shadow"></div>
+                        </div>
+
+                        <div
+                            class="rounded-lg border border-primary/10 p-4 shadow-md backdrop-blur-sm transition duration-300 hover:scale-[1.02] hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-1 text-base font-semibold text-black dark:text-white">
+                                Bonhoeffer College locatie Van der Waalslaan
+                                <span class="text-xs font-medium text-zinc-400"> • 4years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                I completed my secondary education at Bonhoeffer College, where I got my MAVO diploma in 4 years. During the COVID
+                                period I spent a lot of time behind my computer doing online classes. That’s also when I found out I really
+                                enjoyed working with computers — I started taking them apart and putting them back together. That curiosity was
+                                the start of my journey into software development.
+                            </p>
+                        </div>
+                    </div>
+
+                    <!-- ITEM 2 -->
+                    <div class="relative pl-12">
+                        <div class="absolute top-2 left-4 -translate-x-1/2">
+                            <div class="absolute -z-10 h-10 w-10 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600/35 blur-lg"></div>
+                            <div class="h-4 w-4 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600 shadow"></div>
+                        </div>
+
+                        <div
+                            class="rounded-lg border border-primary/10 p-4 shadow-md backdrop-blur-sm transition duration-300 hover:scale-[1.02] hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-1 text-base font-semibold text-black dark:text-white">
+                                MBO Software Developer Study Enschede
+                                <span class="text-xs font-medium text-zinc-400"> • 2years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                This is really where it all began for me. For the intake I had to make a small console app in C#. Back then I had
+                                no clue what I was doing — I had to use things I’d never even heard of before. Still, I managed to build it and I
+                                was proud of myself. The study was more focused on game development, but after my first internship at C-o-d-e-s in
+                                Almelo, where I learned Laravel and Vue.js, I knew web development was the path I wanted to take.
+                            </p>
+                        </div>
+                    </div>
+
+                    <!-- ITEM 3 -->
+                    <div class="relative pl-12">
+                        <div class="absolute top-2 left-4 -translate-x-1/2">
+                            <div class="absolute -z-10 h-10 w-10 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600/35 blur-lg"></div>
+                            <div class="h-4 w-4 -translate-x-1/2 -translate-y-1/2 rounded-full bg-purple-600 shadow"></div>
+                        </div>
+
+                        <div
+                            class="rounded-lg border border-primary/10 p-4 shadow-md backdrop-blur-sm transition duration-300 hover:scale-[1.02] hover:border-primary hover:bg-primary/20"
+                        >
+                            <h3 class="mb-1 text-base font-semibold text-black dark:text-white">
+                                MBO Software Developer Study Hengelo
+                                <span class="text-xs font-medium text-zinc-400"> • 2years</span>
+                            </h3>
+                            <p class="text-sm text-zinc-700 dark:text-zinc-300">
+                                After 2 years in Enschede I decided to switch from game development to web development. Together with 4 friends I
+                                moved to the Hengelo location. At first it wasn’t that easy, since we jumped straight into the third year and
+                                missed a lot of things the others already learned. But with some hard work we caught up quickly. Right now I’m in
+                                my fourth year doing my internship at Cube in Oldenzaal, where I’m learning a lot every day. After this, it’s time
+                                for my final exams.
+                            </p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+
+</template>
