+++ routes/web.php
@@ -2,10 +2,14 @@
 
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
+use App\Http\Controllers\ContactController;
 
-Route::get('/', function () {
-    return Inertia::render('Welcome');
-})->name('home');
+//Route::get('/', function () {
+//    return Inertia::render('Welcome');
+//})->name('home');
 Route::get('/', function () {
     return Inertia::render('Home');
 })->name('home');
+Route::post('/contact', [ContactController::class, 'send'])
+    ->middleware('throttle:contact')
+    ->name('contact.send');
